refactoring del codice, aggiungendo un controller che si occupa dell'intero processo di validazione dell'input
